Moved parsers from 'br/presenter/parser' to 'br/parsing'.
