+++ brjs-sdk/sdk/libs/javascript/br-parsing/br-lib.conf
@@ -1,0 +1,1 @@
+requirePrefix: br/parsing
