+++ brjs-sdk/sdk/libs/javascript/br-parsing/src/zh.properties
@@ -1,0 +1,3 @@
+br.parsing.number.formatting.multiplier.k=1000
+br.parsing.number.formatting.multiplier.m=1000000
+br.parsing.number.formatting.multiplier.b=1000000000
