+++ brjs-sdk/sdk/libs/javascript/br-parsing/test-unit/resources/en.properties
@@ -1,0 +1,3 @@
+br.parsing.number.formatting.multiplier.l=50
+br.parsing.number.formatting.multiplier.c=100
+br.parsing.number.formatting.multiplier.x=10
