--- brjs-sdk/sdk/libs/javascript/br-presenter/src/br/presenter/parser/zh.properties
@@ -1,3 +1,0 @@
-br.presenter.parser.number.formatting.multiplier.k=1000
-br.presenter.parser.number.formatting.multiplier.m=1000000
-br.presenter.parser.number.formatting.multiplier.b=1000000000
