+++ brjs-sdk/sdk/libs/javascript/br-presenter/test-unit/resources/en.properties
@@ -1,5 +1,1 @@
-br.presenter.parser.number.formatting.multiplier.l=50
-br.presenter.parser.number.formatting.multiplier.c=100
-br.presenter.parser.number.formatting.multiplier.x=10
-
 br.presenter.i18ntesttoken=i18nErrorMessage
